Introduce separate activity to ask for permission
