+++ classical-components-demo/idcard-scanner/src/main/res/layout/activity_scanner.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".ScannerActivity"
+        tools:ignore="HardcodedText,PrivateResource"
+        tools:background="#000"
+        >
+
+    <io.scanbot.sdk.ui.camera.ScanbotCameraXView
+            android:id="@+id/cameraView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            />
+
+    <Button android:id="@+id/flashButton"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:text="Flash"
+            />
+
+    <TextView
+            android:id="@+id/resultTextView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="60dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/shutterButton"
+            android:background="#D31C1A1A"
+            android:textColor="#FFF"
+            android:textStyle="bold"
+            android:padding="10dp"
+            tools:text="Result: status=SuccessFound"
+            />
+
+    <io.scanbot.sdk.ui.camera.ShutterButton
+            android:id="@+id/shutterButton"
+            android:layout_width="@dimen/take_picture_size"
+            android:layout_height="@dimen/take_picture_size"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginBottom="28dp"
+            app:shutterButtonDefaultMode="Manual"
+            android:elevation="6dp"
+            />
+</androidx.constraintlayout.widget.ConstraintLayout>
